+++ source/common/router/BUILD
@@ -280,7 +280,6 @@ envoy_cc_library(
     ],
     hdrs = [
         "router.h",
-        "router_filter_interface.h",
         "upstream_request.h",
     ],
     deps = [
@@ -300,6 +299,7 @@ envoy_cc_library(
         "//envoy/http:filter_interface",
         "//envoy/http:stateful_session_interface",
         "//envoy/local_info:local_info_interface",
+        "//envoy/router:router_filter_interface",
         "//envoy/router:shadow_writer_interface",
         "//envoy/runtime:runtime_interface",
         "//envoy/server:factory_context_interface",
